backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
web3==6.11.1
python-dotenv==1.0.0
pyzbar==0.1.9
opencv-python==4.8.0.76
pillow==10.0.1
qrcode==7.4.2
numpy==1.24.3
transformers==4.35.2
torch==2.1.0
gunicorn==21.2.0
python-jose==3.3.0 